增加input validate

+++ app/Http/Controllers/MessageController.php
@@ -18,11 +18,30 @@ class MessageController extends Controller
         $user = Auth::user();
         $input = $request->all();
 
+        $rules = [
+            'classId' => 'required|alpha_num|max:12',
+            'fatherId' => 'nullable|integer',
+            'message' => 'required|max:300',
+        ];
+
+        $messages = [
+            'required' => ':attribute 為必填欄位',
+            'alpha_dash' => ':attribute 請勿輸入特殊字元',
+            'integer' => ':attribute 只限輸入數字',
+            'max'=> ':attribute 不可超過 :max 個字',
+        ];
+
+        $validator = Validator::make($input, $rules, $messages);
+
+        if ($validator->fails()) {
+            return redirect("/class/$classId")->withErrors($validator)->withInput();
+        }
+
         Message::create([
             'classId' => $input['classId'],
             'fatherId' => $input['fatherId'],
             'userName' => $user->name,
-            'message' => $input['message'],
+            'message' => htmlspecialchars($input['message']),
         ]);
     }
 
@@ -48,6 +67,23 @@ class MessageController extends Controller
         $user = Auth::user();
         $input = $request->all();
 
+        $rules = [
+            'id' => 'integer',
+            'message' => 'required|max:300',
+        ];
+
+        $messages = [
+            'required' => ':attribute 為必填欄位',
+            'integer' => ':attribute 只限輸入數字',
+            'max'=> ':attribute 不可超過 :max 個字',
+        ];
+
+        $validator = Validator::make($input, $rules, $messages);
+
+        if ($validator->fails()) {
+            return redirect("/class/$classId")->withErrors($validator)->withInput();
+        }
+
         Message::where('id', $id)->where('userName', $user->name)->update(['message', $input['message']]); 
     }
 
@@ -60,6 +96,20 @@ class MessageController extends Controller
     {
         $user = Auth::user();
         $input = $request->all();
+
+        $rules = [
+            'id' => 'integer',
+        ];
+
+        $messages = [
+            'integer' => ':attribute 只限輸入數字',
+        ];
+
+        $validator = Validator::make($input, $rules, $messages);
+
+        if ($validator->fails()) {
+            return redirect("/class/$classId")->withErrors($validator);
+        }
 
         Message::where('id', $id)->where('userName', $user->name)->delete();
     }
